Authorization and Authentication done using jwt

+++ controller/api/v1/posts_api.js
@@ -19,22 +19,27 @@ module.exports.index = async function (req, res) {
 };
 
 module.exports.destroy = async function (req, res) {
-	try {
-		let post = await Post.findById(req.params.id);
-		post.remove();
-		await Comment.deleteMany({
-			post: req.params.id
-		});
-		return res.status(200).json({
-			message: "post and Comments deleted",
-		})
-		
-	} catch (error) {
-		console.log('*******', error);
-		return res.sttaus(500).json({
-			message: "internal Server Error",
-		});
-	}
+  try {
+    let post = await Post.findById(req.params.id);
+    if (post.user == req.user.id) {
+      post.remove();
+      await Comment.deleteMany({
+        post: req.params.id,
+      });
+      return res.status(200).json({
+        message: 'post and Comments deleted',
+      });
+    } else {
+      return res.json(401, {
+        message: 'you cannot delete this post',
+      });
+    }
+  } catch (error) {
+    console.log('*******', error);
+    return res.sttaus(500).json({
+      message: 'internal Server Error',
+    });
+  }
 };
 
 // module.exports.destroy = async function (req, res) {
